+++ conda_env_cpu.yaml
@@ -1,0 +1,14 @@
+#name: conda_env_name
+
+channels:
+  - pytorch
+  - conda-forge
+  - defaults
+
+dependencies:
+  - python=3.8
+  - pip
+  - notebook
+  - pytorch
+  - torchvision
+  - torchaudio
