+++ configs/callbacks/default_callbacks.yaml
@@ -1,0 +1,16 @@
+model_checkpoint:
+    _target_: pytorch_lightning.callbacks.ModelCheckpoint
+    monitor: "val_acc"      # name of the logged metric which determines when model is improving
+    save_top_k: 2           # save k best models (determined by above metric)
+    save_last: True         # additionaly always save model from last epoch
+    mode: "max"             # can be "max" or "min"
+    dirpath: 'checkpoints/'
+    filename: '{epoch:02d}'
+
+
+early_stopping:
+    _target_: pytorch_lightning.callbacks.EarlyStopping
+    monitor: "val_acc"      # name of the logged metric which determines when model is improving
+    patience: 100           # how many epochs of not improving until training stops
+    mode: "max"             # can be "max" or "min"
+    min_delta: 0.0          # minimum change in the monitored metric needed to qualify as an improvement
