+++ configs/callbacks/mnist_callbacks.yaml
