+++ configs/callbacks/wandb_callbacks.yaml
