+++ configs/config.yaml
@@ -1,0 +1,33 @@
+# @package _global_
+
+# specify here default training configuration
+defaults:
+    - trainer: default_trainer.yaml
+    - model: mnist_model.yaml
+    - datamodule: mnist_datamodule.yaml
+    - callbacks: default_callbacks.yaml  # set this to null if you don't want to use callbacks
+    - logger: null  # set logger here or use command line (e.g. `python train.py logger=wandb`)
+
+    # add this to enable color logging
+    # - override hydra/hydra_logging: colorlog
+    # - override hydra/job_logging: colorlog
+
+
+# path to original working directory (the directory that `train.py` was executed from in command line)
+# hydra hijacks working directory by changing it to the current log directory,
+# so it's useful to have path to original working directory as a special variable
+# read more here: https://hydra.cc/docs/next/tutorials/basic/running_your_app/working_directory
+original_work_dir: ${hydra:runtime.cwd}
+
+
+# path to folder with data
+data_dir: ${original_work_dir}/data/
+
+
+# output paths for hydra logs
+hydra:
+    run:
+        dir: logs/runs/${now:%Y-%m-%d}/${now:%H-%M-%S}
+    sweep:
+        dir: logs/multiruns/${now:%Y-%m-%d_%H-%M-%S}
+        subdir: ${hydra.job.num}
