+++ configs/datamodule/mnist_datamodule.yaml
