+++ configs/logger/csv_logger.yaml
@@ -1,0 +1,6 @@
+# CSVLogger built in PyTorch Lightning
+
+csv_logger:
+    _target_: pytorch_lightning.loggers.csv_logs.CSVLogger
+    save_dir: "."
+    name: "csv_logger/"
