+++ configs/logger/many_loggers.yaml
@@ -1,0 +1,8 @@
+# train with many loggers at once
+
+defaults:
+    - csv_logger.yaml
+    - wandb.yaml
+    # - neptune.yaml
+    # - comet.yaml
+    # - tensorboard.yaml
