+++ configs/logger/tensorboard.yaml
@@ -1,0 +1,6 @@
+# TensorBoard
+
+tensorboard:
+    _target_: pytorch_lightning.loggers.tensorboard.TensorboardLogger
+    save_dir: "tensorboard/"
+    name: "default"
