+++ configs/logger/wandb.yaml
@@ -1,0 +1,9 @@
+# https://wandb.ai (Weights&Biases)
+
+wandb:
+    _target_: pytorch_lightning.loggers.wandb.WandbLogger
+    project: "env_tests"
+    # entity: ""  # set to name of your wandb team or just remove it
+    # offline: False  # set True to store all logs only locally
+    job_type: "train"
+    save_dir: "."
