+++ configs/model/mnist_model.yaml
