+++ configs/trainer/debug_trainer.yaml
@@ -1,0 +1,18 @@
+# trainer args for debugging model
+
+_target_: pytorch_lightning.Trainer
+gpus: 0  # set -1 to train on all GPUs abailable, set 0 to train on CPU only
+# auto_select_gpus: True
+
+gradient_clip_val: 0.5
+
+# fast_dev_run: True
+limit_train_batches: 1.0
+limit_val_batches: 1.0
+limit_test_batches: 1.0
+val_check_interval: 1.0
+profiler: "simple"  # use profiler to print execution time profiling after training ends
+
+progress_bar_refresh_rate: 10
+weights_summary: "full"
+default_root_dir: "lightning_logs/"
