--- project/configs/logger/all_loggers.yaml
@@ -1,7 +1,0 @@
-# Train with many loggers at once
-defaults:
-    - csv_logger.yaml
-    - wandb.yaml
-    - neptune.yaml
-#    - comet.yaml
-#    - tensorboard.yaml
