--- project/configs/logger/comet.yaml
@@ -1,5 +1,0 @@
-# Comet logger config
-comet:
-    _target_: pytorch_lightning.loggers.comet.CometLogger
-    api_key: ???
-    project_name: "project_template_test"
