--- project/configs/logger/wandb.yaml
@@ -1,8 +1,0 @@
-# Weights&Biases logger config
-wandb:
-    _target_: pytorch_lightning.loggers.wandb.WandbLogger
-    project: "env_tests"
-#    entity: ""                  # set to name of your wandb team or just remove it
-    offline: False              # set True to store all logs only locally
-    job_type: "train"
-    save_dir: "."
