--- project/configs/seeds/default_seeds.yaml
@@ -1,1 +1,0 @@
-pytorch_seed: 12345
