+++ requirements.txt
@@ -1,11 +1,28 @@
+# ----- pytorch ----- #
 torch>=1.7.1
 torchvision>=0.8.2
 torchaudio>=0.7.2
 pytorch-lightning>=1.1.6
+
+# ----- hydra ----- #
 # hydra-core>=1.0.5
-hydra-core>=1.1.0.dev2
+hydra-core>=1.1.0.dev4
+hydra_colorlog>=1.0.0
+hydra-optuna-sweeper>=0.9.0rc2
+
+# ----- loggers ----- #
 wandb>=0.10.15
-neptune-client>=0.5.0
+# neptune-client
+# comet_ml
+# mlflow
+
+# ----- random ----- #
+rich>=9.12.2
 scikit-learn>=0.24.0
 pandas>=1.2.0
-hydra_colorlog>=1.0.0
+
+# ----- linter ----- #
+# black>=20.8b1
+# flake8>=3.8.4
+# pylint>=2.7.1
+# isort>=5.7.0
