+++ src/transforms/mnist_transforms.py
@@ -1,0 +1,10 @@
+from torchvision import transforms
+
+
+mnist_train_transforms = transforms.Compose(
+    [transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))]
+)
+
+mnist_test_transforms = transforms.Compose(
+    [transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))]
+)
